3d commit

+++ src/test/java/avdeyev/vik/MyTest.java
@@ -8,7 +8,7 @@ public class MyTest {
     @Test
 public void someTest(){
         Calculator calculator = new Calculator();
-        int result = calculator.add(10,5);
+        double result = calculator.add(10,5);
         Assertions.assertEquals(15,result);
     }
